Add Hashie::Extensions::DeepFind
This extension adds a `#deep_find` (aliased as `#deep_detect`) and
a `#deep_find_all` (aliased as `#deep_select`) method to any Hash-like
object. These methods perform a depth-first search on the object and its
values and return either the first occurrence (for `#deep_find`) or an
array of all occurrences (for `#deep_find_all`) within the nested
structure of the hash. They work for nested Hash-like objects and nested
Enumerables.

[Closes #156]