Add failing spec for #69
I've confirmed that this is a current bug. The problem is that the from
parameter is redefined each time it occurs, via the lines at
`trash.rb:31` through `trash.rb:34`. If you run pry just before that
block and run just the test at `trash_spec:147`, you'll see that the
`#config=` method is defined once for `value_a` and once for `value_b`,
thus overwriting the fetcher for `value_a`.

I can tell what the bug is, but I'm having trouble wrapping my brain
around how to fix it. Thoughts anyone?