Add ruby-2.2.0 and ruby-2.2.1 to the Travis config
There's an issue that presents in Ruby 2.2 versions that causes a spec
to fail. I'm introducing these versions to Travis so the specs will
properly fail, since the way the specs are written isn't something
that's particularly unlikely with using method accessors.

The reason I'm adding a bunch of versions is to see if we can bisect the
version in which the error started occurring.

See #273