Retain Mash's original key with indifferent access
This is an attempt at fixing the issue of Mash's keys being converted
to strings upon assignment. Since we want Mash to behave like a Hash,
this is undesireable; however, many gems downstream depend on the
indifferent access that is achieved through casting all keys to strings.

I had to touch a few tests that were looking specifically for a certain
type of key (i.e. relying on the fact that keys are cast). They should
now be semantically correct.

There is one interesting thing that pops out from this. Checking for
equality between Mashes (and Hashes with Mashes) now fails unless the
keys are equivalent. For example:

```ruby
Hashie::Mash.new(p: 'test') == Hashie::Mash.new('p' => 'test') #=> false
```

I also added a test for issue #246, which is solved by this PR.

I _believe_ that this should perform on par with the current
implementation. I'm going to work on a benchmark to double-check, since
a performance regression is less desireable than the fix that this
gives.

Fixes #196
Fixes #246