Add documentation about circular coercion
The `coerce_key` method in the Coercion extension is a class-level
method. As such, using circular coercion is not possible without using
a coercion proc/lambda, since the coercion in the first type will not
have the second type defined at load time.

This can be worked around through the use of a coercion proc. I have
added docs about this to the README, as well as tests that demonstrate
the behavior.

I don't see this as a bug, since doing this raises a Ruby-level error
that makes sense.

The only way I can think of to make the non-working demonstration work
that I can think of would make the DSL a lot noisier (i.e. make all of
the "intos" use procs/lambdas to yield the class name/structure that
you'd like. Something like `coerce_key :model, to: -> { Model }`, which
then grabs that returned `Model` and is applied. I think this is a
reduction in user experience, so I chose not to pursue it.

Fixes #115