type system include builder
broke out the builtin Hashie type coercions into a system that is able
to load different type coercion systems in a configurable way. The new
CoercionSystemIncludeBuilder creates anonymous modules with `included`
methods that will correctly include the overall coercion modules and the
sub type system override modules into the base class. It also is reused
in the `Dash::Coercion` class, so that we do not need to repeat any code
there. Default behavior of running `include
Hashie::Extensions::Coercion` retains the same behavior to make this a
non-breaking change.