Add a logging layer to address common issues
One of the most common issues we have reported is when a `Hashie::Mash`
attempts to override a built-in method and fails to do so. To prevent
this from being an issue in the future, I added a logging layer
accessible at `Hashie.logger` that is user-configurable. This should
help to report the error to the user before it becomes a problem.

The logging layer is a barebones implementation right now. It defaults
to a standard `Logger.new($stdout)` but can be overridden by setting
`Hashie.logger = MyAwesomeLogger.new`. Ideally, we can expand this layer
and its usage to make the user experience better for Hashie.

Closes #361