Allow Mash subclasses to disable warnings
Since we are transitively used as a dependency in many projects, we
should have given the ability to toggle this behavior off. The logging
feature is more of a "help people get started with Mash" feature. If
you're using Hashie in a library, it's likely that you already know the
tradeoffs of attempting to override methods on the object.

To use this feature, you only have to subclass Mash and then call the
class method:

```ruby
class KeyStore < Hashie::Mash
  disable_warnings
end
```