Remove auto-load for `Hashie.logger`
The auto-load is never triggered because there is no `Hashie::Logger`
constant to trigger it. We need to `require` the file in the top level
`hashie.rb` and anywhere else that uses it (in case someone just loads a
single file).

At the moment, this is only used in Mash, so we're covered.