Allow the dig method on StrictKeyAccess hashes
Currently, if you create a Hash with StrictKeyAccess mixed in, then you do
`some_strict_hash.dig 'alpha', 'beta'`, you will get a `DefaultError`
because ruby 2.4.0 tries to set the default somewhere in the dig mechanism
(it's not actually clear to me where, I looked at the source briefly).

This patch allows dig to work in a StrictKeyAccess hash just as it does
in a default hash.

My use case for this patch is that I want to have a hash that raises KeyError
by default on `[]` access for a key that does not exist, but I also want to
be able to explicitly allow `nil` when I use `dig`. In my particular case,
I've also mixed in `DeepFetch`, so I can use `deep_fetch` to distinguish
cases where I want to allow `nil` from cases where I don't.'