Add yard-doctest and fix all known broken examples
To test all of our documentation examples, you can now run:

    $ bundle exec yard doctest

There are a few problems in the library that are causing issues with the
Coercion specs, so those are disabled for now. I'm going to see if I can
figure out the problem and fix it so we don't have to skip those tests.