Refactor PropertyTranslation for readability
The original implementation of PropertyTranslation was heavily reliant
longer methods that added a lot of cognitive load for understanding what
should be happening at any given point. This refactor gives names for all
of the steps that happen at each step to make it easier to reason about
bugs and additions.