Ensure we return nils for unset keys in Dash
In 0bd18eeb, we stopped setting any `nil` values on the Dash in order to
prevent co-dependent properties from throwing errors. This is a problem
because it means that when you explicitly set `nil` as a value for
a property, it's not set on the underlying hash and you will not get it
back if you call `#to_h`.

By explicitly setting a nil value on the result of `#to_h` for unset
keys, we can avoid this issue and return what the called expects.