Exclude tests from the gem release
When you're installing a gem in a production environment, you want it to
install as fast it can. One of the ways you can speed up the
installation of the gem is by making it smaller. We currently ship the
test suite with the gem, increasing the size of the built gem
significantly.

By not shipping the test suite, we can shrink the size of the gem by
38%. Below are the measurements I took for that statement.

**The size of the gem with the test suite**

$ du -b hashie-4.0.1.gem
80384	hashie-4.0.1.gem

**The size of the gem without the test suite**

$ du -b hashie-4.0.1.gem
50176	hashie-4.0.1.gem