Don't warn when setting most affixed keys (#500)
Due to how we have implemented the bang/underbang/query behavior within
Mash, setting keys that have those affixes in them actually allow
overwriting the behavior of those affixes. As such, we shouldn't warn
when setting a key that matches those patterns.

When it comes to setter-like keys, I believe we still _do_ want to warn
for two reasons:

1. Trying to access the key via method access is a syntax error. Ruby
   expects any method ending in `=` to be a 2+-arity method due to the
   infix notation of setter methods. This is unexpected behavior unless
   you're very familiar with Ruby parsing.
2. You can still retrieve the key via the normal `Hash#[]` reader, but
   it prevents setting a similar key without the equal sign. You can see
   this in the test about setters. I'd say that is unexpected and
   surprising behavior.

Because of these two gotchas, I think we should still warn in cases
where you try to set a key that looks like a setter.