Add logger as a runtime dependency
Starting in Ruby 3.4, requiring `logger` without it being in the lockfile will emit a warning.
In Ruby 3.5 it will fail.

In development here, `faraday` already depends on `logger` so there are no warnings. But, obviously you can use hashie without it.

```
]$ bundle exec rspec
/home/user/.rbenv/versions/3.4.1/lib/ruby/gems/3.4.0/gems/hashie-4.1.0/lib/hashie/logger.rb:1: warning: logger was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add logger to your Gemfile or gemspec to silence this warning.
```